+++ frontend/src/app/trees/admin/trees-admin.service.spec.ts
@@ -1,7 +1,6 @@
-import { TestBed, async, inject } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { TreesAdminService } from './trees-admin.service';
-import { forest } from '../_mocks/forest.mock';
-import { FormBuilder, Validators } from '@angular/forms';
+import { FormBuilder} from '@angular/forms';
 
 describe('TreesAdminService', () => {
   let service: TreesAdminService;
