+++ frontend/src/app/trees/forests/tree-guidelines/tree-cutting-dates/tree-cutting-dates.component.spec.ts
@@ -1,11 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FilterPipe } from '../../../../_pipes/filter.pipe';
-import { forest } from '../../../_mocks/forest.mock';
+import { forest } from '../../../../_mocks/forest.mock';
 import { TreeCuttingDatesComponent } from './tree-cutting-dates.component';
-import { Pipe, PipeTransform } from '@angular/core';
 import { LineBreakFormatterPipe } from '../../../../_pipes/line-break-formatter.pipe';
-import * as sinon from 'sinon';
 import { TreeDistrictsUtilService } from '../tree-districts-util.service';
 import { WindowRef } from '../../../../_services/native-window.service';
 import { MarkdownService } from 'ngx-md';
