+++ frontend/src/app/trees/forests/tree-guidelines/tree-cutting/tree-cutting-rules/tree-cutting.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { TreeCuttingRulesComponent } from './tree-cutting-rules.component';
-import { forest } from '../../../../_mocks/forest.mock';
+import { forest } from '../../../../../_mocks/forest.mock';
 
 describe('TreeCuttingRulesComponent', () => {
   let component: TreeCuttingRulesComponent;
