+++ frontend/src/app/trees/forests/tree-guidelines/tree-rules/tree-rules.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FilterPipe } from '../../../../_pipes/filter.pipe';
-import { forest } from '../../../_mocks/forest.mock';
+import { forest } from '../../../../_mocks/forest.mock';
 import { TreeRulesComponent } from './tree-rules.component';
 import { LineBreakFormatterPipe } from '../../../../_pipes/line-break-formatter.pipe';
 import { SpacesToDashesPipe } from '../../../../_pipes/spaces-to-dashes.pipe';
