+++ frontend/src/app/trees/forests/tree-guidelines/tree-species/tree-species.component.spec.ts
@@ -3,7 +3,7 @@ import { Component, CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from '@angular/co
 import { FilterPipe } from '../../../../_pipes/filter.pipe';
 import { SpacesToDashesPipe } from '../../../../_pipes/spaces-to-dashes.pipe';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { forest } from '../../../_mocks/forest.mock';
+import { forest } from '../../../../_mocks/forest.mock';
 
 import { ForestService } from '../../../_services/forest.service';
 import { TreeSpeciesComponent } from './tree-species.component';
